backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.8
sentry-sdk==1.34.0
requests==2.31.0
prometheus-client==0.20.0
psutil==5.9.8